test(quotes): split extraction and append-extraction API tests
